Refactors app for modularity and maintainability
Breaks down the monolithic `app.js` into smaller, more manageable modules,
improving code organization and reusability.

This change introduces dedicated modules for:
- UI handling (tabs, buttons, resize)
- Data loading
- Feature-specific logic (polls, candidates)
- Plotly utilities

This improves the project's structure and makes it easier to maintain
and extend in the future.
