+++ plotly-standalone/index.html
@@ -128,6 +128,6 @@
 		</div>
 	</footer>
 
-	<script src="assets/js/app.js"></script>
+	<script type="module" src="assets/js/app.js"></script>
 </body>
 </html>
